# WIP: Make Node.WithLatests() take a slice instead of a map