Then do the additions as a merge rather than repeated adds, to avoid memory allocations.

It appears that `WithLatests` is not something we do a lot in the benchmarks, but probably more in the probes. The follow-up change to `propagateLatest` gives us a healthy boost when rendering.

~Also a drive-by fix to `processWithContainerNameRenderer`.~

Benchmarks:
```
benchmark                            old ns/op      new ns/op      delta
BenchmarkRenderList-2                665307761      527805245      -20.67%
BenchmarkRenderHosts-2               413642784      408523260      -1.24%
BenchmarkRenderControllers-2         364129636      344651406      -5.35%
BenchmarkRenderPods-2                341522570      314960182      -7.78%
BenchmarkRenderContainers-2          204275436      193586083      -5.23%
BenchmarkRenderProcesses-2           129523723      122403160      -5.50%
BenchmarkRenderProcessNames-2        143726808      138044775      -3.95%

benchmark                            old allocs     new allocs     delta
BenchmarkRenderList-2                801654         803633         +0.25%
BenchmarkRenderHosts-2               605887         605869         -0.00%
BenchmarkRenderControllers-2         429942         429789         -0.04%
BenchmarkRenderPods-2                383866         385131         +0.33%
BenchmarkRenderContainers-2          253898         255118         +0.48%
BenchmarkRenderProcesses-2           119585         120188         +0.50%
BenchmarkRenderProcessNames-2        163417         163872         +0.28%

benchmark                            old bytes     new bytes     delta
BenchmarkRenderList-2                91623684      88391116      -3.53%
BenchmarkRenderHosts-2               65873664      62455445      -5.19%
BenchmarkRenderControllers-2         47176061      43729286      -7.31%
BenchmarkRenderPods-2                41874317      38607350      -7.80%
BenchmarkRenderContainers-2          28799258      25524452      -11.37%
BenchmarkRenderProcesses-2           13984811      14002312      +0.13%
BenchmarkRenderProcessNames-2        17930731      17981694      +0.28%
```