@bboreham 2018-04-12
We don't know how small the change to probe performance is; we don't have any benchmarks.
We could try it out...

The readability change comes from gofmt aligning map entries and not parameters, right?