@rade 2018-04-12
> We don't know how small the change to probe performance is; we don't have any benchmarks.

True.

> We could try it out...

If there is a significant performance improvement, I would be swayed.

> The readability change comes from gofmt aligning map entries and not parameters, right?

Yes. And also from the fact that the representation of an association as a flat list (there is a name for that, which I've forgotten) is rather non-idiomatic golang.