@bboreham 2018-07-17
Based on profiling some real probes, I wrote a benchmark to mimic typical operation of this code when reporting on Kubernetes objects.

Results before (best of 5):
```
BenchmarkReporter 	     500	   2671920 ns/op	  841966 B/op	    9290 allocs/op
```

After:
```
BenchmarkReporter 	     500	   2260789 ns/op	  516960 B/op	    9142 allocs/op
```