@bboreham 2018-07-18
Replaced by #3268